+++ CHANGELOG.md
@@ -68,6 +68,7 @@ All notable changes to mainline dht will be documented in this file.
 - `Server::handle_request()` signature change, to avoid circular dependency on `Rpc`.
 - Make `DefaultServer` properties public.
 - Trait `Server` needs to implement `Clone`, but no longer needs to implement `Sync`.
+- `DhtBuilder` is not consuming, thanks to `Config` being `Clone`.
 
 ##  [4.2.0](https://github.com/pubky/mainline/compare/v4.1.0...v4.2.0) - 2024-12-13